Move Request class to utils, add generate_token method
