+++ app/api/views.py
@@ -1,79 +1,9 @@
-import hashlib
 import jwt
-from jwt import ExpiredSignatureError, InvalidTokenError
-import datetime
-
 from flask_restful import Resource
-from flask_restful.reqparse import RequestParser
-from app.db.models import Bucketlist, Item, User
 from sqlalchemy.exc import IntegrityError
+from app.db.models import Bucketlist, Item, User
 from app.app import session, SECRET_KEY
-# Moving 13- 16 out of view
-
-
-class Request(RequestParser):
-    def __init__(self):
-        """
-        Utility class used for performing generic manipulations on incoming request objects.
-        Initialised with all the arguments to be used by the view classes.
-
-        """
-        super().__init__()
-        self.add_argument('name', location='form')
-        self.add_argument('done', location='form')
-        self.add_argument('username', location='form')
-        self.add_argument('password', location='form')
-        self.add_argument('token', location='headers')
-        self.add_argument('page', location='args')
-        self.add_argument('limit', location='args')
-
-    def set_password(self, password):
-        """
-        Method that salts a password string using the SECRET_KEY;
-        to return a sha256 hashed string.
-        """
-
-        salted_password = password + SECRET_KEY
-
-        hashed_password = hashlib.sha256(salted_password.encode()).hexdigest()
-        return hashed_password
-
-    def is_authenticated(self):
-        """
-        Method used to identify the user of an incoming request via the token;
-        returns True if user is valid.
-        """
-        try:
-            token = self.parse_args()['token']
-            user_data = jwt.decode(token, key=SECRET_KEY)
-            self.current_user = session.query(User).filter_by(username=user_data['username']).first()
-            return True
-        # AttributeError separated
-        except (AttributeError, ExpiredSignatureError, InvalidTokenError):
-            return False
-
-    def save(self, obj):
-        session.add(obj)
-        session.commit()
-
-    def remove(self, obj):
-        session.delete(obj)
-        session.commit()
-
-    def paginated(self, obj_list):
-        """ Method used to paginate results in an object list """
-        self.page = 1 if not self.parse_args().get('page', 1) else int(self.parse_args().get('page'))
-        self.limit = 20 if not self.parse_args().get('limit', 20) else int(self.parse_args().get('limit'))
-
-        paginator_obj = obj_list.paginate(1, self.limit)
-        self.total_pages = paginator_obj.pages
-
-        # If the requested page number is out of bounds, return the last page.
-        if self.page > self.total_pages:
-            self.page = self.total_pages
-
-        paginated_list = obj_list.paginate(self.page, self.limit)
-        return paginated_list.items
+from . utils import Request
 
 
 class Register(Request, Resource):
@@ -98,10 +28,7 @@ class Register(Request, Resource):
                         password=secure_password)
         self.save(new_user)
 
-        # Token payload is encoded with the new user's username and an expiry period.
-        payload = {'username': new_user.username,
-                   "exp": datetime.datetime.utcnow() + datetime.timedelta(seconds=600)}
-        auth_token = jwt.encode(payload, key=SECRET_KEY).decode()
+        auth_token = self.generate_token(new_user.username)
 
         return {'auth_token': auth_token}, 200
 
@@ -122,9 +49,7 @@ class Login(Request, Resource):
             return 'Check username and password', 400
 
         # Token payload is encoded with username and expiry date
-        payload = {'username': user.username,
-                   "exp": datetime.datetime.utcnow() + datetime.timedelta(seconds=600)}
-        auth_token = jwt.encode(payload, key=SECRET_KEY).decode()
+        auth_token = self.generate_token(user.username)
 
         return {'auth_token': auth_token}, 200
